consumer/requirements.txt
pika==1.1.0
mysql-connector-python==8.0.22